+++ public/js/i18n/en_US.js
@@ -202,6 +202,7 @@ gc2i18n = {
         "Copy all properties from another layer": "Copy all properties from another layer",
         "Layer properties copied": "Layer properties copied",
         "Enable filtering": "Enable filtering",
+        "Disable filtering": "Disable filtering",
         "Show in conflict": "Show in conflict",
         "Make link": "Make link",
         "Show in click info": "Show in click info",
@@ -230,6 +231,8 @@ gc2i18n = {
         "Append": "Append",
         "Write tag names into the field 'Tags'. Finish the tag name with an 'Enter' key stroke. If 'Append' is checked, new tags are appended to existing one. Append is only availably when mulitple layers are selected.": "Write tag names into the field 'Tags'. Finish the tag name with an 'Enter' key stroke. If 'Append' is checked, new tags are appended to existing one. Append is only availably when mulitple layers are selected.",
 
-        "Change between WMS and Tile Cache": "Change between single WMS and Tile Cache"
+        "Change between WMS and Tile Cache": "Change between single WMS and Tile Cache",
+
+        "Structure": "Structure"
     }
 };